Changes so we no longer require administrator rights.
The program used to write to HKEY_CURRENT_CONFIG, which since Vista needs
admin rights to be able to write. However, this had the advantage that all
config was saved in one place, for all users on the machine. However, the
admin requirement caused confusion, so now we write to HKEY_CURRENT_USER,
which is *per user*. Each user will have their own config.

On startup if we can't read the current user's config, we try to read an old
config from HKEY_CURRENT_CONFIG, and if that fails, use compiled-in defaults.
All writes go to the per-user registry now though. The sneaky can still maintain
a global config if they dare to use the registry editor though, and don't write
user configs.